==> 1-posm-vs-postcode.my.ipynb <==
import pandas as pd
import numpy as np
import src.utils

# Pos Malaysia data vs. postcode.my data

We have two separate data files containing postcode data; one from Pos Malaysia and one scraped from postcode.my. They don't map cleanly, so let's investigate.

# Import files, check the hash to ensure no changes have been made.
POSM = "./data/Poscode - Lat Long.xlsx"
POSTCODEMY = "./data/dep_poskod_geo_web.csv"
assert src.utils.check_hash(POSM)
assert src.utils.check_hash(POSTCODEMY)

df_posm_import = pd.read_excel(
    POSM, 
    sheet_name="Sheet2", 
    converters={"POSTCODE": src.utils.postcode_str}
) # Pos Malaysia data
df_postcodemy_import = pd.read_csv(
    POSTCODEMY, 
    converters={"postcode": src.utils.postcode_str}
) # postcode.my data

print("PosM data:")
print(df_posm_import.info())
print("-----")
print("postcode.my data:")
print(df_postcodemy_import.info())

# Keep only the columns which are important to us.
df_posm = df_posm_import["POSTCODE"].rename("postcode").drop_duplicates()
df_postcodemy = df_postcodemy_import["postcode"].drop_duplicates()

print("Pos Malaysia data:")
display(df_posm.head())
print(df_posm.shape)
print("-----")
print("postcode.my data:")
display(df_postcodemy.head())
print(df_postcodemy.shape)


Here are the:

- **Postcodes in postcode.my, which are *not* in PosM** - we will be using this.
- **Postcodes in PosM, which are *not* in postcode.my** - for informational purposes only.

not_in_posm = df_postcodemy[~df_postcodemy.isin(df_posm)].sort_values()
not_in_postcodemy = df_posm[~df_posm.isin(df_postcodemy)].sort_values()

print("In postcode.my, NOT in PosM:")
print(not_in_posm)
print("-----")
print("In PosM, NOT in postcode.my:")
print(not_in_postcodemy)

# Closest valid PosM postcode

Let's call each postcode.my postcode which isn't in PosM an **invalid** postcode. We want to match it to the closest **valid** postcode in the PosM dataset.

We'll use a simple heuristic: for each postcode, let's call its first two digits the **region**. E.g. 50480 --> 50, 07893 --> 07

For each invalid postcode, find the set of all valid postcodes with the same region. Then, pick the valid postcode whose last 3 digits is closest to the invalid postcode's last 3 digits, when treated as an integer.

Example: Invalid postcode - **31900**. Valid postcodes with region 31 - **{31000, 31800, 31890, 31920}**. Then the closest valid postcode is **31890** (difference of 10).

# Run the heuristic.
results = []
for pc in not_in_posm:
    similar_mask = df_posm.apply(lambda x: x[:2] == pc[:2]) # Same first 2 digits (inefficient, but small dataset...)
    similar = df_posm[similar_mask].to_list()
    if len(similar) == 0:
        raise ValueError(f"{pc} has no Pos Malaysia postcode with the same first 2 digits!")
    nearest = min(similar, key=lambda x: abs(int(x[2:]) - int(pc[2:])))
    results.append((pc, nearest))
df = pd.DataFrame(results, columns=["postcode.my", "closest_posm"])
display(df)

# TODO: check if these coordinates are close enough

# Export files

df.to_csv("./output/postcode.my-to-posm.csv", sep="|", index=False)
