==> 3-distance-matrix.ipynb <==
import csv
import itertools

import openpyxl
import pandas as pd

import src.utils


# Import in df_pc from Notebook 2
df_pc = pd.read_pickle("./output/posm-postcodes-full.pkl")

# Get just the master postcodes. Sort by region for a nicer matrix later.
df_mpc = df_pc.drop(columns="postcode").drop_duplicates()
df_mpc = df_mpc.sort_values(by=["ppv_region", "master_postcode"])

display(df_mpc)

 # Distance Matrix

 Generate a distance matrix: the haversine distance between any two master postcodes.
 Postcodes in different `ppv_region`s will be blank (NA) as we can't travel across borders.

 Since this is quite expensive computation-wise, we will also generate a csv which
 stores all the valid master postcode pairings and coordinates. This will be used for for feeding into Google Maps API.

# A list of dicts containing all valid master postcode pairings, to be converted to csv later.
mpc_pairs = []

# Generate distance matrix.
pc_labels = [(x, y) for x, y in zip(df_mpc["ppv_region"], df_mpc["master_postcode"])]
master_lat_lon = {x: (y, z) 
    for x, y, z 
    in zip(df_mpc["master_postcode"], df_mpc["lat"], df_mpc["lon"])
}

# dataframe_matrix
df_mat = pd.DataFrame(
    pd.NA,
    index=pd.MultiIndex.from_tuples(pc_labels),
    columns=pd.MultiIndex.from_tuples(pc_labels),
)

# Diagonals are zero distance
for tup in pc_labels:
    df_mat.loc[tup, tup] = 0

# Off-diagonals, calculate the haversine distance.
for tup1, tup2 in itertools.combinations(pc_labels, 2):
    ppv1, mpc1 = tup1
    ppv2, mpc2 = tup2
    if ppv1 == ppv2:
        if mpc1 == mpc2:
            df_mat.loc[tup1, tup2] = 0
        else:
            lat1, lon1 = master_lat_lon[mpc1]
            lat2, lon2 = master_lat_lon[mpc2]
            distance = src.utils.haversine_km(lat1, lon1, lat2, lon2)
            df_mat.loc[tup1, tup2] = distance
            df_mat.loc[tup2, tup1] = distance  # reflect across the diagonal
            # This is a valid pair, so add it to the list.
            mpc_pairs.append(
                {
                    "ppv_region": ppv1,
                    "master_postcode_1": mpc1,
                    "lat_1": lat1,
                    "lon_1": lon1,
                    "master_postcode_2": mpc2,
                    "lat_2": lat2,
                    "lon_2": lon2,
                }
            )

display(df_mat)

# Export files

# Save the distance matrix.
NAME = "distance-matrix"

# Excel for easy human reading, since it preserves MultiIndex.
EXCEL_NAME = f"./output/{NAME}.xlsx"
df_mat.to_excel(EXCEL_NAME)
workbook = openpyxl.load_workbook(EXCEL_NAME)
workbook["Sheet1"].freeze_panes = "C3"  # Slightly nicer to scroll through.
workbook.save(EXCEL_NAME)

# Save as pickle for future notebooks.
df_mat.to_pickle(f"./output/{NAME}.pkl")


# Save the master postcode pairs for later notebooks:
with open("./output/master-postcode-pairs.csv", mode="w", newline="") as fp:
    writer = csv.DictWriter(fp, fieldnames=mpc_pairs[0].keys(), delimiter="|")
    writer.writeheader()
    for pair in mpc_pairs:
        writer.writerow(pair)

# Estimated Google Maps API cost

num = len(mpc_pairs)
print(f"There are a total of {num} valid master postcode pairs.")
print(f"Total cost in Maps API: USD{num * 0.005:2g}.")
print("-----")
print(f"You can find the data output files in the `output` folder.")
