==> 4-google-maps-api.ipynb <==
import csv
import json
from pprint import pprint

import googlemaps
import numpy as np
import pandas as pd

import src.utils

# Google Maps Distance Matrix API

We now have every valid pairing of master postcodes, and their coordinates. As a reminder:

- Master postcodes in one ppv_region cannot reach another ppv_region (due to state travel restrictions).
- Selangor, KL and Putrajaya are considered the same ppv_region.
- Each other state is its own ppv_region.

We will use the Google Maps Distance Matrix API to obtain the time required to travel from one valid master postcode to another.

As the API returns a different result at different times, it is best to perform the all requests at the same time, preferable at night when roads are empty. This way, we can avoid traffic jams messing up our time between postcodes.

# Import the list of all master postcode pairings from previous notebook.
with open("./output/master-postcode-pairs.csv") as fp:
    mpc_pairs = list(csv.DictReader(fp, delimiter="|"))
# Note: The lat and lon are strings instead of floats, but that's ok as we need strings for the API anyways.

pprint(mpc_pairs[:3])

# For testing purposes, it is best to hit the API with one ppv_region at a time.
# Let's separate this by ppv_region:
mpc_pairs_by_region = dict()
for mpc_pair in mpc_pairs:
    ppv = mpc_pair["ppv_region"]
    if ppv not in mpc_pairs_by_region.keys():
        mpc_pairs_by_region[ppv] = []
    mpc_pairs_by_region[ppv].append(mpc_pair)

print("Example: Kedah (first 3)")
pprint(mpc_pairs_by_region["Kedah"][:3])
print("-----")
print("Number of valid pairings by state:")
pprint({k: len(v) for k, v in mpc_pairs_by_region.items()})


# The ppv_regions we can use for API testing.
print(mpc_pairs_by_region.keys())

state = "Terengganu"
mpc_pairs_api = mpc_pairs_by_region[state] # <-- This is what is fed into the API. Change as you please.

SKIP = True # Change this to False to use the Maps API.
if not SKIP:
    gmaps = googlemaps.Client(key=src.utils.get_api_key())
    responses = []  # All the responses we get from API.
    ts = src.utils.now()
    # Store the responses in a json file, inside ./responses folder.
    with open(f"./responses/distance-matrix-{ts}-{state}.json", mode="a") as fp:
        for item in mpc_pairs_api:
            # coordinates separated by comma, no spaces
            origin = f"{item['lat_1']},{item['lon_1']}"
            destination = f"{item['lat_2']},{item['lon_2']}"
            try:
                response = gmaps.distance_matrix(origin, destination)
                # pprint(response)  # for debugging
            except Exception as exc:
                pprint(exc)
            responses.append(response)
        json.dump(responses, fp)
    print("First response:")
    pprint(responses[0])
    print("-----")
    print(f"Total responses: {len(responses)}")
    ok_resp = sum(1 for r in responses if r["status"] == "OK")
    print(f"OK responses: {ok_resp}")


# Augment the mpc_pairs_api dictionaries with the duration returned by API.
state = "Terengganu"
NAME = "20210611-001855-Terengganu"
mpc_pairs_api = mpc_pairs_by_region[state] # <-- This is what is fed into the API. Change as you please.

with open(f"./responses/final/distance-matrix-{NAME}.json") as fp: # Use this line for sample
    responses = json.load(fp)

# Add the corresponding duration into the mpc_pairs dictionaries
try:
    for item, resp in zip(mpc_pairs_api, responses):
        element = resp["rows"][0]["elements"][0]
        if element["status"] == "ZERO_RESULTS":
            item["time_sec"] = "inf"
        else:
            item["time_sec"] = element["duration"]["value"]
except KeyError as exc:
    print(resp)
    raise exc

pprint(mpc_pairs_api[:3])


# Export files

assert len(mpc_pairs_api) == len(responses)  # Check all responses are here
with open(f"./output/master-postcode-pairs-{NAME}.csv", mode="w", newline="") as fp:
    writer = csv.DictWriter(fp, fieldnames=mpc_pairs_api[0].keys(), delimiter="|")
    writer.writeheader()
    for pair in mpc_pairs_api:
        writer.writerow(pair)
