==> 6-ppv-in-master-postcode.ipynb <==
import numpy as np
import pandas as pd
import src.utils

df = pd.read_pickle("./output/posm-postcodes-full.pkl")

config = {
    "sheet_name": "static",
    "usecols": ["code", "state", "district", "lat", "lon", "ppv"],
    "index_col": 0,
    "converters": {"state": lambda s: s.replace(".", "")},  # W.P. Kuala Lumpur -> WP Kuala Lumpur
}
ppv_public = pd.read_excel("./data/sensitive/dep_ppv.xlsx", **config)
ppv_private = pd.read_excel("./data/sensitive/dep_ppv_phcorp.xlsx", **config)



# Data Cleaning

for ppv in [ppv_public, ppv_private]:
    ppv["ppv_region"] = ppv["state"].apply(
        lambda x: "Selangor/KL/Putrajaya"
        if x in ["Selangor", "WP Kuala Lumpur", "WP Putrajaya"]
        else x
    )

display(ppv_public)
display(ppv_private)

# We only want the master postcodes for this analysis.
df = df.drop(columns="postcode").drop_duplicates().set_index("master_postcode")

def closest_mpc(x: pd.Series):
    temp = df.loc[:, ["lat", "lon"]]
    temp["dist"] = temp.apply(
        lambda y: src.utils.haversine_km(x["lat"], x["lon"], y["lat"], y["lon"]),
        axis="columns"
    )
    nearest = temp.sort_values("dist", ascending=True).iloc[0]
    return (nearest.name, nearest["lat"], nearest["lon"], nearest["dist"])

for df_ppv in [ppv_public, ppv_private]:
    df_ppv[["nearest_master", "master_lat", "master_lon", "dist"]] = \
        df_ppv.apply(closest_mpc, axis="columns", result_type="expand")
    display(df_ppv)   

print("Public PPV distance from master postcode:")
display(ppv_public["dist"].describe())
print("-----")
print("Private PPV distance from master postcode:")
display(ppv_private["dist"].describe())


gb_public = ppv_public.groupby("nearest_master").count().apply(max, axis="columns").to_dict()
gb_private = ppv_private.groupby("nearest_master").count().apply(max, axis="columns").to_dict()

df[["num_ppv_public", "num_ppv_private"]] = (
    df.apply(
        lambda x: (gb_public.get(x.name, 0), gb_private.get(x.name, 0)),
        axis="columns",
        result_type="expand",
    )
)
df["total_num_ppv"] = df[["num_ppv_public", "num_ppv_private"]].apply(sum, axis="columns")
display(df)

# Which master postcode has no PPV?
mpc_no_ppv = df[df["total_num_ppv"] == 0]
display(mpc_no_ppv)

ppv = pd.concat([ppv_public, ppv_private])

def closest_ppv_in_region(x):
    """Given a Series from the master postcode df, what is the closest ppv in the same region?"""
    mpc_region = x["ppv_region"]
    temp = ppv[ppv["ppv_region"] == mpc_region] # All ppvs in this region
    temp["dist"] = temp.apply(
        lambda y: src.utils.haversine_km(x["lat"], x["lon"], y["lat"], y["lon"]),
        axis="columns"
    )
    nearest = temp.sort_values("dist", ascending=True).iloc[0]
    return (nearest.name, nearest["lat"], nearest["lon"], nearest["dist"])

df[["closest_ppv", "ppv_lat", "ppv_lon", "ppv_dist"]] = \
    df.apply(closest_ppv_in_region, axis="columns", result_type="expand")

# display(df)
print("-----")
print("MPCs with no linked PPV only - now they have the closest PPV in the region")
mpc_no_ppv = df[df.index.isin(mpc_no_ppv.index)]
display(mpc_no_ppv)


# Export file

df.to_excel("./output/master-postcodes-and-ppvs.xlsx")
